+++ test/test_parser_logic.gd
@@ -13,7 +13,8 @@ func _create_doc_payload(content = [], blocks = []):
 			"type": "content",
 			"content": content
 		}],
-		"blocks": blocks
+		"blocks": blocks,
+		"links": {},
 	}
 
 func test_condition_single_var():
@@ -1212,5 +1213,3 @@ func test_empty_block():
 		"conditions": null,
 	}])
 	assert_eq_deep(result, expected)
-
-